Changelog:
> Updated authentication routine, now an "apikey" variable is sent on the header for authentication purposes.
The apikey is hardcoded, but also can be passed as parameter in the initialization function in case Cocos Capital update it.
apikey is found on the main.js file of the webpage. A method for scrapping and getting that value will be developed in the future, if needed.

> Updated validation list of instrument types and subtypes due to a modification on the server and removal of an endpoint.

> Updated buying power validation on CI settlement.

> Other minor changes

All available methods are working fine now.